// tsrpc_server/src/module/role/model/RoleEnum.ts
/*
 * @Author: dgflash
 * @Date: 2022-01-26 14:14:34
 * @LastEditors: dgflash
 * @LastEditTime: 2022-08-18 15:00:24
 */

/** 角色属性类型 */
export enum RoleAttributeType {
    /** 当前生命值 */
    hp = "hp",
    /** 生命最大值 */
    hpMax = "hpMax",
    /** 移动速度 */
    speed = "speed",
    /** 攻击力 */
    ad = "ad",
    /** 魔力 */
    ap = "ap"
}